pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "n", email = "n"},
]
dependencies = [
    "instructor>=0.4.2",
    "loguru>=0.7.2",
    "aiofiles>=23.2.1",
    "fastapi>=0.105.0",
    "uvicorn>=0.24.0.post1",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "requests[socks]>=2.31.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "n"}

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "pytest>=7.4.3",
]
